2.PVE模式城外放点【Done】 3.防御塔血量由100改为225【Done】
5.灵盾上限不变，效果由被攻击时触发改为自动回复，每五秒可回复灵盾等级一半的盾值。
解决噬魂导致闪电链特效残留bug【Done】
所有人补偿100礼物点数
解决程序多次读取指令bug【Done】
血条格数优化【Done】
凋零制作完成【Done】
凋零生成时间

+++ Assets/Scripts/SkillSystem/SkillBase.cs
@@ -22,7 +22,7 @@ public abstract class SkillBase : ScriptableObject
     public Sprite icon;
     [Header("是否是Buff[暂未使用]")] public bool isBuff;
     
-    public int cd;//冷却时间
+    public float cd;//冷却时间
     [HideInInspector]public float timer = 0;
 
     [Header("可使用次数,-1表示无限次")]
